+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  * Add extension `VK_EXT_blend_operation_advanced` (revision 2)
  * Add extension `VK_NV_fragment_coverage_to_color` (revision 1)
  * Add extension `VK_NV_framebuffer_mixed_samples` (revision 1)
+ * Add extension `VK_NV_fill_rectangle` (revision 1)
 
 ## Version 0.19.0, released on 25.06.2017